requirements.txt
Pillow>=8
matplotlib